Move everything to root, delete dashboard
