+++ apps/dashboard/src/routes/modals/LinkDetailModal.svelte
